feature: rename main module to AuthFree
