+++ src/pages/index.ts
@@ -5,7 +5,7 @@ import NotesPage from "../modules/notes/pages/NotesPage";
 import TrashPage from "../modules/notes/pages/TrashPage";
 
 // Public Pages
-import PublicHomePage from "../modules/main/pages/PublicHomePage";
-import DownloadPage from "../modules/main/pages/DownloadPage";
+import PublicHomePage from "../modules/AuthFree/presentation/pages/PublicHomePage";
+import DownloadPage from "../modules/AuthFree/presentation/pages/DownloadPage";
 
 export { HomePage, NotesPage, TrashPage, DownloadPage, PublicHomePage };
